pyproject.toml
[tool.poetry]
name = "auth"
version = "0.0.1"
authors = ["PauLark inc."]
description = "Another one auth service"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
dynaconf = "^3.2.5"
motor = "^3.4.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pyjwt = "^2.8"
email-validator = "^2.2"
fastapi-mail = "^1.4.1"
httpx = "^0.27"
asgi-lifespan = "^2.1.0"
pytest-httpx = "^0.30"
alt-pytest-asyncio = "^0.8.1"
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
types-passlib = "1.7.7.20240327"
types-requests = "2.32.0.20240712"

[tool.mypy]
plugins = ["pydantic.mypy"]
disable_error_code = "valid-type"
show_error_codes = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["dynaconf.*"]
ignore_missing_imports = true

[tool.isort]
profile = "black"
py_version = 312

[tool.flake8]
max_line_length = 79
ignore = ["E203"]

[tool.black]
target_version = ["py312"]
line-length = 79